fix(pulse-core): use f64 for confidence/eval_score, tidy README
serde_json widens f32 to f64 before formatting, so the committed
fixtures showed noise like 0.949999988079071 instead of the clean
0.95 that production JSON (summaries/v3/*.json) and
apps/daily-agent/src/manifest.rs already carry for eval_score.
Regenerated docs/contracts/brief.json accordingly, and split the
contracts README run-on sentence into a shorter paragraph.

+++ libs/pulse-core/src/lib.rs
@@ -44,7 +44,7 @@ pub struct InsightBrief {
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct BriefMeta {
-    pub confidence: Option<f32>,
+    pub confidence: Option<f64>,
     pub category: Option<String>,
 }
 
@@ -59,7 +59,7 @@ pub struct Brief {
     pub article_url: String,
     pub article_title: String,
     pub model: Option<String>,
-    pub eval_score: Option<f32>,
+    pub eval_score: Option<f64>,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
